Add some flags to the compilation, and correct a warning in memtree.
* SConstruct: add -O2 and -Wall to the compilation

* src/memtree.c:
  (print_node): do some casting magic to make the argument compatible with
    the printf format code

+++ src/memtree.c
@@ -824,7 +824,7 @@ print_node(const struct pc_memtree_s *node, int depth)
         printf("null\n");
         return;
     }
-    printf("%s:%d\n", MT_COLOR(node), MT_SIZE(node));
+    printf("%s:%lu\n", MT_COLOR(node), (unsigned long)MT_SIZE(node));
 
     print_node(node->smaller, ++depth);
     print_node(node->larger, depth);
